+++ mackerel-plugin-openldap/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/mackerelio/mackerel-agent-plugins/mackerel-plugin-openldap/lib"
+
+func main() {
+	mpopenldap.Do()
+}