+++ mackerel-plugin-plack/lib/plack.go
@@ -2,7 +2,6 @@ package mpplack
 
 import (
 	"encoding/json"
-	"errors"
 	"flag"
 	"fmt"
 	"io"
@@ -96,25 +95,22 @@ func (p PlackPlugin) parseStats(body io.Reader) (map[string]interface{}, error) 
 		return nil, err
 	}
 
-	stat["busy_workers"], err = strconv.ParseFloat(s.BusyWorkers, 64)
-	if err != nil {
-		return nil, errors.New("cannot get values")
+	if s, err := strconv.ParseFloat(s.BusyWorkers, 64); err == nil {
+		stat["busy_workers"] = s
 	}
 
-	stat["idle_workers"], err = strconv.ParseFloat(s.IdleWorkers, 64)
-	if err != nil {
-		return nil, errors.New("cannot get values")
+	if s, err := strconv.ParseFloat(s.IdleWorkers, 64); err == nil {
+		stat["idle_workers"] = s
 	}
 
-	stat["requests"], err = strconv.ParseUint(s.TotalAccesses, 10, 64)
-	if err != nil {
-		return nil, errors.New("cannot get values")
+	if s, err := strconv.ParseUint(s.TotalAccesses, 10, 64); err == nil {
+		stat["requests"] = s
 	}
 
-	stat["bytes_sent"], err = strconv.ParseUint(s.TotalKbytes, 10, 64)
-	if err != nil {
-		return nil, errors.New("cannot get values")
+	if s, err := strconv.ParseUint(s.TotalKbytes, 10, 64); err == nil {
+		stat["bytes_sent"] = s
 	}
+
 	return stat, nil
 }