+++ mackerel-plugin.go
@@ -60,7 +60,7 @@ func run(args []string) int {
 	return exitOK
 }
 
-const version = "0.29.1"
+const version = "0.31.0"
 
 var gitcommit string